added junit groovy impelemnted eating

+++ src/main/java/pl/sdacademy/animals/Animal.java
@@ -4,5 +4,7 @@ public interface Animal {
 
     boolean isAlive();
 
-    int getWeight();
+    double getWeight();
+
+
 }
